pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ralph-pp"
version = "0.1.0"
description = "Automated orchestration for the Ralph agentic coding workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-slugify>=8.0",
]

[project.scripts]
"ralph++" = "ralph_pp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["ralph_pp"]

[dependency-groups]
dev = ["pytest>=8.0", "pyright>=1.1.380", "ruff>=0.8"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
